long int

+++ holberton.h
@@ -6,6 +6,6 @@
 int _printf(const char *format, ...);
 int _putchar(char c);
 int print_str(char *c);
-int print_numbers(int n);
+int print_numbers(long int n);
 
 #endif
